chore: use ESM config and clean unused vars
