+++ README.md
@@ -166,3 +166,5 @@ yarn build
 - [Next.js Documentation](https://nextjs.org/docs)
 - [Tailwind CSS Documentation](https://tailwindcss.com/docs)
 - [Framer Motion Documentation](https://www.framer.com/motion/)
+
+Deployment: initialize prod pipeline
