+++ src/components/ServicesPageClient.tsx
@@ -394,7 +394,7 @@ export default function ServicesPageClient({ servicesDetail }: ServicesPageClien
       {/* CTA Section */}
       <section
         id="contact" // Ensure this ID exists if used for internal links
-        className="py-20 px-4 md:px-8 bg-blue-600 text-white"
+        className="py-20 px-4 md:px-8 bg-blue-600 text-white scroll-mt-20"
       >
         <div className="max-w-7xl mx-auto text-center">
           <h2 className="text-3xl md:text-4xl font-bold mb-6">
